@@ -4,50 +4,24 @@ import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
 class MkScreenUtilConfig {
-  const MkScreenUtilConfig({
-    this.width = 1080,
-    this.height = 1920,
-    this.allowFontScaling = false,
-  });
+  const MkScreenUtilConfig({this.width = 1080, this.height = 1920, this.allowFontScaling = false});
 
   final double width;
   final double height;
   final bool allowFontScaling;
 }
 
-mixin MkScreenUtilProvider<T extends StatefulWidget> on State<T> {
-  MkScreenUtilConfig get screenConfig;
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) async {
-      assert(screenConfig != null);
-      MkScreenUtil.initialize(
-        context: context,
-        config: screenConfig,
-      );
-    });
-  }
-}
-
 class MkScreenUtil {
   factory MkScreenUtil() => _instance;
 
-  MkScreenUtil._({
-    @required BuildContext context,
-    this.config,
-  }) : mq = MediaQuery.of(context);
+  MkScreenUtil._({@required BuildContext context, this.config}) : mq = MediaQuery.of(context);
 
   static MkScreenUtil initialize({
     @required BuildContext context,
     MkScreenUtilConfig config = const MkScreenUtilConfig(),
   }) {
     assert(config != null);
-    return _instance ??= MkScreenUtil._(
-      context: context,
-      config: config,
-    );
+    return _instance ??= MkScreenUtil._(context: context, config: config);
   }
 
   static MkScreenUtil _instance;
