@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 
-abstract class MkKeyboardProvider {
+abstract class DismissKeyboardProvider {
   BuildContext get context;
 
   void closeKeyboard() => FocusScope.of(context).requestFocus(FocusNode());
