@@ -20,11 +20,11 @@ import 'package:tailor_made/screens/homepage/_views/out_dated.dart';
 import 'package:tailor_made/screens/homepage/_views/rate_limit.dart';
 import 'package:tailor_made/screens/splash/splash.dart';
 import 'package:tailor_made/services/accounts/accounts.dart';
-import 'package:tailor_made/utils/mk_choice_dialog.dart';
-import 'package:tailor_made/utils/mk_navigate.dart';
 import 'package:tailor_made/utils/mk_phone.dart';
-import 'package:tailor_made/utils/mk_status_bar.dart';
+import 'package:tailor_made/utils/ui/mk_choice_dialog.dart';
+import 'package:tailor_made/utils/ui/mk_status_bar.dart';
 import 'package:tailor_made/widgets/_partials/mk_loading_spinner.dart';
+import 'package:tailor_made/wrappers/mk_navigate.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({
@@ -60,7 +60,7 @@ class HomePage extends StatelessWidget {
                 converter: (store) => HomeViewModel(store),
                 builder: (
                   BuildContext context,
-                  DispatchFunction dispatcher,
+                  DispatchFunction dispatch,
                   HomeViewModel vm,
                 ) {
                   if (vm.isLoading) {
@@ -91,10 +91,10 @@ class HomePage extends StatelessWidget {
                   if (vm.isWarning && vm.hasSkipedPremium == false) {
                     return RateLimitPage(
                       onSignUp: () {
-                        dispatcher(OnPremiumSignUp(payload: vm.account));
+                        dispatch(OnPremiumSignUp(payload: vm.account));
                       },
                       onSkipedPremium: () {
-                        dispatcher(const OnSkipedPremium());
+                        dispatch(const OnSkipedPremium());
                       },
                     );
                   }
@@ -152,7 +152,7 @@ class _Body extends StatelessWidget {
           shouldSendRating: vm.shouldSendRating,
           onLogout: () async {
             await Accounts.di().signout();
-            StoreProvider.of<AppState>(context).dispatcher(
+            StoreProvider.of<AppState>(context).dispatch(
               const OnLogoutAction(),
             );
             await Navigator.pushAndRemoveUntil<void>(
