@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:rebloc/rebloc.dart';
+import 'package:tailor_made/providers/dispatch_provider.dart';
 import 'package:tailor_made/rebloc/app_state.dart';
 import 'package:tailor_made/rebloc/jobs/actions.dart';
 import 'package:tailor_made/rebloc/jobs/view_model.dart';
 import 'package:tailor_made/screens/jobs/_partials/jobs_filter_button.dart';
 import 'package:tailor_made/screens/jobs/_partials/jobs_list.dart';
 import 'package:tailor_made/screens/jobs/jobs_create.dart';
-import 'package:tailor_made/utils/mk_dispatch_provider.dart';
-import 'package:tailor_made/utils/mk_navigate.dart';
 import 'package:tailor_made/widgets/_partials/mk_app_bar.dart';
 import 'package:tailor_made/widgets/_partials/mk_close_button.dart';
 import 'package:tailor_made/widgets/_partials/mk_loading_spinner.dart';
 import 'package:tailor_made/widgets/theme_provider.dart';
+import 'package:tailor_made/wrappers/mk_navigate.dart';
 
 class JobsPage extends StatelessWidget {
   const JobsPage({Key key}) : super(key: key);
@@ -22,7 +22,7 @@ class JobsPage extends StatelessWidget {
       converter: (store) => JobsViewModel(store),
       builder: (
         BuildContext context,
-        DispatchFunction dispatcher,
+        DispatchFunction dispatch,
         JobsViewModel vm,
       ) {
         return WillPopScope(
@@ -45,10 +45,9 @@ class JobsPage extends StatelessWidget {
             floatingActionButton: FloatingActionButton(
               child: const Icon(Icons.library_add),
               onPressed: () {
-                MkNavigate(
-                  context,
+                Navigator.of(context).push<void>(MkNavigate.slideIn(
                   JobsCreatePage(contacts: vm.contacts),
-                );
+                ));
               },
             ),
           ),
@@ -81,7 +80,7 @@ class _AppBar extends StatefulWidget implements PreferredSizeWidget {
   Size get preferredSize => Size.fromHeight(kToolbarHeight);
 }
 
-class _AppBarState extends State<_AppBar> with MkDispatchProvider<AppState> {
+class _AppBarState extends State<_AppBar> with DispatchProvider<AppState> {
   bool _isSearching = false;
 
   @override
